fix ascyng tasks when pulling

+++ src/main/java/com/ksig/vcs_cli/async/PullItemsTask.java
@@ -2,6 +2,7 @@ package com.ksig.vcs_cli.async;
 
 import com.ksig.vcs_cli.http.BackendRestClient;
 import com.ksig.vcs_cli.models.SyncItemView;
+import com.ksig.vcs_cli.models.enums.ItemType;
 import org.apache.hc.client5.http.classic.methods.HttpGet;
 import org.apache.hc.core5.net.URIBuilder;
 
@@ -10,55 +11,72 @@ import java.net.URI;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.concurrent.Callable;
 
-public class PullItemsTask implements Runnable {
+public class PullItemsTask implements Callable<Void> {
     private final SyncItemView itemToSync;
     private final BackendRestClient backendRestClient;
-    private final URIBuilder baseURL;
-    public  PullItemsTask(SyncItemView itemToSync, BackendRestClient backendRestClient, URIBuilder baseURL) {
+    private final URI baseURL;
+
+    public PullItemsTask(SyncItemView itemToSync, BackendRestClient backendRestClient, URI baseURL) {
         this.itemToSync = itemToSync;
         this.backendRestClient = backendRestClient;
         this.baseURL = baseURL;
     }
 
     @Override
-    public void run() {
+    public Void call() {
         switch (itemToSync.getStatus()) {
             case DELETED_REMOTE: deleteLocalFile(itemToSync); break;
-            case NEW_REMOTE: downloadRemoteFile(itemToSync);
+            case NEW_REMOTE: downloadRemoteFile(itemToSync); break;
             case MODIFIED_REMOTE: handleModifiedRemoteFile(itemToSync); break;
         }
+        return null;
     }
 
     private void deleteLocalFile(SyncItemView itemToSync) {
         try {
             Files.delete(itemToSync.getLocalPath());
+            System.out.println("Deleted local " + itemToSync.getPath());
         } catch (IOException e) {
-            System.err.println("Error deleting local file " + itemToSync.getPath());
+            System.err.println("Error deleting local file " + itemToSync.getPath() + " Local path: " + itemToSync.getLocalPath() );
+            e.printStackTrace();
         }
-        System.out.println("Deleted local file " + itemToSync.getPath());
     }
 
     private void downloadRemoteFile(SyncItemView itemToSync) {
         try {
-            URI uri = baseURL.appendPath("content").appendPath(itemToSync.getStorageKey()).build();
+            if (itemToSync.getItemType() == ItemType.DIRECTORY) {
+                Files.createDirectories(itemToSync.getLocalPath());
+                System.out.println("Created local directory " + itemToSync.getPath());
+                return;
+            }
+
+            URI uri = new URIBuilder(baseURL).appendPath("content").appendPath(itemToSync.getStorageKey()).build();
             HttpGet httpGet = new HttpGet(uri);
-            backendRestClient.downloadFile(httpGet, itemToSync.getLocalPath());
+            backendRestClient.downloadFileWithFileName(httpGet, itemToSync.getLocalPath());
+            System.out.println("Downloaded remote file " + itemToSync.getPath());
         } catch (Exception e) {
-            System.err.println("Error downloading remote file " + itemToSync.getPath());
+            System.err.println("Error downloading remote file " + itemToSync.getPath() + " Local path: " + itemToSync.getLocalPath() );
+            e.printStackTrace();
         }
-        System.out.println("Downloaded remote file " + itemToSync.getPath());
     }
 
     private void handleModifiedRemoteFile(SyncItemView itemToSync) {
         try {
+            if (itemToSync.getItemType() == ItemType.DIRECTORY) {
+                System.out.println("Ignored modified status for directory " + itemToSync.getPath());
+                return;
+            }
+
             Files.delete(itemToSync.getLocalPath());
-            URI uri = baseURL.appendPath("content").appendPath(itemToSync.getStorageKey()).build();
+            URI uri = new URIBuilder(baseURL).appendPath("content").appendPath(itemToSync.getStorageKey()).build();
             HttpGet httpGet = new HttpGet(uri);
-            backendRestClient.downloadFile(httpGet, itemToSync.getLocalPath());
+            backendRestClient.downloadFileWithFileName(httpGet, itemToSync.getLocalPath());
+            System.out.println("Changed remote modified file " + itemToSync.getPath());
         } catch (Exception e) {
-            System.err.println("Error handling remote file " + itemToSync.getPath());
+            System.err.println("Error handling remote file " + itemToSync.getPath() + " Local path: " + itemToSync.getLocalPath() );
+            e.printStackTrace();
         }
-        System.out.println("Changed remote modified file " + itemToSync.getPath());
     }
 }
